medical_communication_app_ver2/requirements.txt
fastapi
uvicorn
jinja2
requests
numpy
sentence-transformers
python-multipart
aiofiles
torch
transformers
sentencepiece
whisper
pyannote.audio